Vectors in second graphic corrected
Had to change the values passed in to degrees and radius distance passed in as 'x' and 'y' respectively. For some reason the polar axes sizes are not of the specified width.

At one point they were of the correct size on the page. Perhaps there is a property/key value 'pgfpolarplotsset' specifically to ensure the axes are of the correct width. Otherwise, due to pdflatex page formatting or internal pgfplots parameters are modifying the graphics on the page so that they are forced to fit on the same page.

Considering that these two grpahics are technically notation/formulas information I might move them to the opening section of the document. It is not absolutely necessary that this property of the Complex numbers is outlined in this part of the document.